refactor: app 모듈에서 남은 feature:provide-picker 코드 분리
