+++ .gitignore
@@ -137,3 +137,6 @@ fabric.properties
 ### AndroidStudio Patch ###
 
 !/gradle/wrapper/gradle-wrapper.jar
+
+#Module
+**/*UseCaseImpl.kt
