+++ app/src/main/AndroidManifest.xml
@@ -1,24 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
-
-    <uses-feature
-        android:name="android.hardware.camera"
-        android:required="false" />
-
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-    <uses-permission android:name="android.permission.CAMERA" />
-    <uses-permission android:name="${providerPermission}" />
     <queries>
         <package android:name="com.skt.skaf.l001mtm091"/>
         <package android:name="com.skt.tmap.ku"/>
         <intent>
             <action android:name="android.media.action.IMAGE_CAPTURE" />
         </intent>
-        <provider android:authorities="@string/admin_authority" />
     </queries>
     <application
         android:name=".BaseApplication"
