+++ feature/provider-picker/src/main/AndroidManifest.xml
@@ -1,4 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
+    <uses-permission android:name="${providerPermission}" />
+    <queries>
+        <provider android:authorities="@string/admin_authority" />
+    </queries>
 </manifest>
