+++ feature/provider-picker/src/main/java/com/dhkim139/feature/providerpicker/ProviderPickerModule.kt
@@ -1,0 +1,21 @@
+package com.dhkim139.feature.providerpicker
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ProviderPickerModule {
+
+    @Provides
+    @Singleton
+    fun provideGetImagesPageUseCase(@ApplicationContext context: Context): GetImagesPageUseCase{
+        return GetImagesPageUseCaseImpl(context)
+    }
+}
