+++ feature/provider-picker/src/main/java/com/dhkim139/feature/providerpicker/ProviderPickerViewModel.kt
@@ -10,7 +10,6 @@ import androidx.paging.PagingState
 import androidx.paging.cachedIn
 import com.dhkim139.feature.providerpicker.model.PickerImage
 import com.dhkim139.feature.providerpicker.model.PickerImage.Companion.toPickerImage
-import com.wheretogo.domain.usecase.util.GetImagesPageUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
