+++ feature/provider-picker/src/main/res/values/config.xml
