+++ feature/provider-picker/src/release/res/values/config.xml
